+++ Cargo.toml
@@ -3,15 +3,40 @@ name = "milkshake-terminal"
 version = "0.0.0"
 edition = "2021"
 
+[[bin]]
+name = "milkshake-terminal"
+path = "src/main.rs"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[features]
+default = []
+fontconfig = ["font-kit/source-fontconfig", "font-kit/loader-freetype"]
+
 [dependencies]
 base64 = { version = "0.22.1", default-features = false, features = ["std"] }
-bevy = { version = "0.15.0-rc.2", default-features = false, features = ["bevy_pbr", "bevy_text", "bevy_ui", "multi_threaded", "tonemapping_luts", "wayland"] }
+bevy = { version = "0.15.0-rc.2", default-features = false, features = [
+    "bevy_pbr",
+    "bevy_text",
+    "bevy_ui",
+    "multi_threaded",
+    "tonemapping_luts",
+    "wayland",
+] }
 compact_str = { version = "0.8.0", default-features = false }
 crossbeam-channel = { version = "0.5.13", default-features = false }
-fontconfig = { version = "0.9.0", default-features = false, features = ["dlopen"] }
-image = { version = "0.25.4", default-features = false, features = ["jpeg", "png"] }
+font-kit = { version = "0.14.2" }
+image = { version = "0.25.4", default-features = false, features = [
+    "jpeg",
+    "png",
+] }
 libc = { version = "0.2.161", default-features = false }
-rustix = { version = "0.38.37", default-features = false, features = ["libc", "process", "std"] }
+rustix = { version = "0.38.37", default-features = false, features = [
+    "libc",
+    "process",
+    "std",
+] }
 rustix-openpty = { version = "0.1.1", default-features = false }
 vte = { version = "0.13.0", default-features = false }
 
@@ -20,3 +45,7 @@ opt-level = 2
 
 [profile.dev.package."*"]
 opt-level = 3
+
+[patch.crates-io]
+# https://github.com/servo/pathfinder/issues/571
+pathfinder_simd = { git = "https://github.com/theoparis/pathfinder" }
